✨ omit grapher explorers from explorer-views-and-charts index
